Extract constructor from trait to implementations
- See zendframwork/zf2#7676
