+++ src/Validator/ValidatorChainTrait.php
@@ -21,26 +21,6 @@ trait ValidatorChainTrait
     protected $storage;
 
     /**
-     * Construct the validation chain
-     *
-     * Retrieves validators from session storage and attaches them.
-     *
-     * @param StorageInterface $storage
-     */
-    public function __construct(StorageInterface $storage)
-    {
-        parent::__construct();
-
-        $this->storage = $storage;
-        $validators = $storage->getMetadata('_VALID');
-        if ($validators) {
-            foreach ($validators as $validator => $data) {
-                $this->attachValidator('session.validate', [new $validator($data), 'isValid'], 1);
-            }
-        }
-    }
-
-    /**
      * Retrieve session storage object
      *
      * @return StorageInterface
